requirements.txt
# Core dependencies
deepgram-sdk>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pyaudio
sounddevice
openai

# Audio processing
sounddevice>=0.4.6
numpy>=1.24.0
